time_calculator
This program takes a Starting Time, and additional hours and minutes, and the optional argument of Day. Returns a new time where additional hours and minutes are added to  the Starting Time and returns the Ending Time including including updated AM/PM, the amount of days later if applicable, and the Ending Day if the optional argument was given.